Dandole detalles a la lista de favoritos

+++ frontend/src/components/ListaFavorito/ListaFavorito.jsx
@@ -9,10 +9,10 @@ const ListaFavorito = () => {
   useEffect(() => {
     const fetchFavorites = async () => {
       try {
-        const response = await axios.get("http://localhost:3001/fav");
+        const response = await axios.get('http://localhost:3001/fav');
         setFavorites(response.data);
       } catch (error) {
-        console.error("Error al obtener la lista de favoritos:", error);
+        console.error('Error al obtener la lista de favoritos:', error);
       }
     };
 
@@ -20,23 +20,28 @@ const ListaFavorito = () => {
   }, []);
 
   const removeFromFavorites = async (id) => {
-    try {
-      await axios.delete(`http://localhost:3001/fav/${id}`);
-      setFavorites((prevFavorites) =>
-        prevFavorites.filter((favorite) => favorite.id !== id)
-      );
-      console.log("Eliminé el producto de favoritos:", id);
-      Swal.fire(
-        "Eliminado de favoritos",
-        "El producto se ha eliminado de favoritos",
-        "success"
-      );
-    } catch (error) {
-      Swal.fire(
-        "Error",
-        "No se pudo eliminar el producto de favoritos",
-        "error"
-      );
+    const answer = await Swal.fire({
+      title: '¿Estás seguro?',
+      text: '¿Deseas eliminar este producto de favoritos?',
+      icon: 'question',
+      showCancelButton: true,
+      confirmButtonText: 'Sí, eliminar',
+      cancelButtonText: 'No',
+    });
+
+    if (answer.isConfirmed) {
+      try {
+        await axios.delete(`http://localhost:3001/fav/${id}`);
+        setFavorites((prevFavorites) =>
+          prevFavorites.filter((favorite) => favorite.id !== id)
+        );
+        console.log('Eliminé el producto de favoritos:', id);
+        Swal.fire('Eliminado de favoritos', 'El producto se ha eliminado de favoritos', 'success');
+      } catch (error) {
+        Swal.fire('Error', 'No se pudo eliminar el producto de favoritos', 'error');
+      }
+    } else {
+      console.log('No se eliminó el producto de favoritos:', id);
     }
   };
 
@@ -47,7 +52,6 @@ const ListaFavorito = () => {
       prevFavorites.map((favorite) => ({ ...favorite, selected: newSelectAll }))
     );
   };
-  
 
   const handleCheckboxChange = (event, id) => {
     const isChecked = event.target.checked;
@@ -59,31 +63,49 @@ const ListaFavorito = () => {
   };
 
   const handleRemoveSelected = async () => {
-    const selectedIds = favorites
-      .filter((favorite) => favorite.selected)
-      .map((favorite) => favorite.id);
+    const selectedFavorites = favorites.filter((favorite) => favorite.selected);
 
-    try {
-      await Promise.all(
-        selectedIds.map((id) => axios.delete(`http://localhost:3001/fav/${id}`))
-      );
-      setFavorites((prevFavorites) =>
-        prevFavorites.filter((favorite) => !selectedIds.includes(favorite.id))
-      );
-      console.log("Eliminé los productos de favoritos:", selectedIds);
-      Swal.fire(
-        "Eliminados de favoritos",
-        "Los productos se han eliminado de favoritos",
-        "success"
-      );
-    } catch (error) {
-      Swal.fire(
-        "Error",
-        "No se pudieron eliminar los productos de favoritos",
-        "error"
-      );
+    if (selectedFavorites.length === 0) {
+      Swal.fire('Seleccionar uno para poder eliminar', '', 'warning');
+      return;
+    }
+
+    const answer = await Swal.fire({
+      title: '¿Estás seguro?',
+      text: '¿Deseas eliminar los productos seleccionados de favoritos?',
+      icon: 'question',
+      showCancelButton: true,
+      confirmButtonText: 'Sí, eliminar',
+      cancelButtonText: 'No',
+    });
+
+    if (answer.isConfirmed) {
+      const selectedIds = selectedFavorites.map((favorite) => favorite.id);
+
+      try {
+        await Promise.all(
+          selectedIds.map((id) => axios.delete(`http://localhost:3001/fav/${id}`))
+        );
+        setFavorites((prevFavorites) =>
+          prevFavorites.filter((favorite) => !selectedIds.includes(favorite.id))
+        );
+        console.log('Eliminé los productos de favoritos:', selectedIds);
+        Swal.fire(
+          'Eliminados de favoritos',
+          'Los productos se han eliminado de favoritos',
+          'success'
+        );
+      } catch (error) {
+        Swal.fire('Error', 'No se pudieron eliminar los productos de favoritos', 'error');
+      }
+    } else {
+      console.log('No se eliminaron los productos de favoritos');
     }
   };
+
+  function redirectToStore() {
+    window.location.href = "/store";
+  }
 
   return (
     <div className="flex justify-center items-center h-screen">
@@ -99,7 +121,7 @@ const ListaFavorito = () => {
                       className={`${
                         selectAll ? 'bg-gray-300' : 'bg-transparent'
                       } border-2 border-gray-300 rounded-md text-black py-2 px-4 opacity-75`}
-                    >
+                      >
                       Seleccionar todo
                     </button>
                   </th>
@@ -113,40 +135,60 @@ const ListaFavorito = () => {
                 </tr>
               </thead>
               <tbody>
-                {favorites.map((favorite, index) => (
-                  <tr key={favorite.id}>
-                    <td className="py-2 px-4 border-b">
-                      <input
-                        type="checkbox"
-                        checked={favorite.selected || false}
-                        onChange={(event) => handleCheckboxChange(event, favorite.id)}
-                      />
-                    </td>
-                    <td className="py-2 px-4 border-b">
-                      <div className="flex items-center">
-                        <img
-                          src={favorite.image}
-                          className="py-2 px-4 w-24 h-24 hover:scale-110"
-                          alt={favorite.name ? favorite.name : favorite.title}
-                        />
-                      </div>
-                    </td>
-                    <td className="py-2 px-4 border-b text-black">
-                      {favorite.name ? favorite.name : favorite.title}
-                    </td>
-                    <td className="py-2 px-4 border-b text-black">{favorite.description}</td>
-                    <td className="py-2 px-4 border-b text-black">$ {favorite.price}</td>
-                    <td className="py-2 px-4 border-b text-black">{favorite.rating}</td>
-                    <td className="py-2 px-4 border-b">
-                      <button
-                        onClick={() => removeFromFavorites(favorite.id)}
-                        className="bg-red-500 text-white rounded-full p-2 hover:bg-red-600"
-                      >
-                        Eliminar
-                      </button>
-                    </td>
-                  </tr>
-                ))}
+              {favorites.length > 0 ? (
+  favorites.map((favorite, index) => (
+    <tr key={favorite.id}>
+      <td className="py-2 px-4 border-b">
+        <input
+          type="checkbox"
+          checked={favorite.selected || false}
+          onChange={(event) => handleCheckboxChange(event, favorite.id)}
+        />
+      </td>
+      <td className="py-2 px-4 border-b">
+        <div className="flex items-center">
+          <img
+            src={favorite.image}
+            className="py-2 px-4 w-24 h-24 hover:scale-110"
+            alt={favorite.name ? favorite.name : favorite.title}
+          />
+        </div>
+      </td>
+      <td className="py-2 px-4 border-b text-black">
+        {favorite.name ? favorite.name : favorite.title}
+      </td>
+      <td className="py-2 px-4 border-b text-black">{favorite.description}</td>
+      <td className="py-2 px-4 border-b text-black">${favorite.price}</td>
+      <td className="py-2 px-4 border-b text-black">{favorite.rating}⭐</td>
+      <td className="py-2 px-4 border-b">
+        <button
+          onClick={() => removeFromFavorites(favorite.id)}
+          className="bg-red-500 text-white rounded-full p-2 hover:bg-red-600"
+        >
+          Eliminar
+        </button>
+      </td> 
+    </tr>
+  ))
+) : (
+<tr>
+  <td className="bg-gray-100 py-8" colspan="7">
+    <div className="flex flex-col items-center">
+      <h1 className="text-4xl text-red-500">No hay favoritos</h1>
+      <button className="mt-4 px-6 py-3 rounded-md text-white bg-red-500 hover:bg-red-600" onClick={redirectToStore}>Agregar productos</button>
+    </div>
+  </td>
+</tr>
+
+
+
+)}
+{/* <tr>
+<td colSpan="7">
+  <h1 style={{ color: 'red', textAlign: 'center' }}>No hay favoritos</h1>
+  <button style={{color: 'red', textAlign:'center'}}>Agregar productos</button>
+</td>
+</tr> */}
               </tbody>
             </table>
           </div>
@@ -165,6 +207,5 @@ const ListaFavorito = () => {
     </div>
   );
 };
-
 
 export default ListaFavorito;
